Add bounds checking functionality
There's no point in attempting to encode invalid latitude or longitude
coordinates. The function's error value is the input, which we're
converting to an informative error message.

One drawback is the fact that encoding will now fail on the first error;
A more flexible approach would be to first check the values, and if the
vector contains errors, to extract them and their positions into the Err. 🤔